+++ src/parser/block/blockquote.ts
@@ -2,7 +2,7 @@ import { BlockquoteParser } from '../block';
 import { union, some, block, validate, rewrite, creator, fmap, open, convert, lazy } from '../../combinator';
 import { defrag, suppress } from '../util';
 import { contentline } from '../source';
-import { autolink } from '../autolink';
+import { autolinkblock } from '../autolinkblock';
 import { parse } from '../api/parse';
 import { html } from 'typed-dom';
 
@@ -32,7 +32,7 @@ const text: BlockquoteParser.TextParser = lazy(() => fmap(
       convert(unindent, text)),
     rewrite(
       some(contentline, opener),
-      convert(unindent, fmap(some(autolink), ns => [html('pre', defrag(ns))]))),
+      convert(unindent, fmap(autolinkblock, ns => [html('pre', defrag(ns))]))),
   ])),
   ns => [html('blockquote', ns)]));