+++ src/parser/block/paragraph/mention/quotation.test.ts
@@ -36,6 +36,10 @@ describe('Unit: parser/block/paragraph/mention/quotation', () => {
       assert.deepStrictEqual(inspect(parser('>a b')), [['<span class="quotation">&gt;a b</span>'], '']);
       assert.deepStrictEqual(inspect(parser('>a b\nc')), [['<span class="quotation">&gt;a b</span>'], 'c']);
       assert.deepStrictEqual(inspect(parser('>a b\n> c')), [['<span class="quotation">&gt;a b<br>&gt; c</span>'], '']);
+      assert.deepStrictEqual(inspect(parser('> >a\n> > b')), [['<span class="quotation">&gt; <a href="?log=a" rel="noopener" class="address" data-level="1">&gt;a</a><br>&gt; &gt; b</span>'], '']);
+      assert.deepStrictEqual(inspect(parser('> >a\n> > b\n> c')), [['<span class="quotation">&gt; <a href="?log=a" rel="noopener" class="address" data-level="1">&gt;a</a><br>&gt; &gt; b<br>&gt; c</span>'], '']);
+      assert.deepStrictEqual(inspect(parser('> > a\n> > b\n> > c')), [['<span class="quotation">&gt; &gt; a<br>&gt; &gt; b<br>&gt; &gt; c</span>'], '']);
+      assert.deepStrictEqual(inspect(parser('> > > a\n> > > b')), [['<span class="quotation">&gt; &gt; &gt; a<br>&gt; &gt; &gt; b</span>'], '']);
     });
 
   });